Abstracted directives module functionality from single file into directory containing directives module, threads module, and lines module. Ported tests over to match. No changes to core functionality.
